requirements.txt
streamlit
together
pandas
numpy
python-docx
PyMuPDF
pytesseract
Pillow
matplotlib
seaborn
tabulate
openpyxl